+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [unreleased]((https://github.com/NodeFactoryIo/vedran/tree/HEAD))
+
+[Full Changelog](https://github.com/NodeFactoryIo/vedran/compare/v.0.1.1...HEAD)
+
+### Added
+- Implement tunnel server in load balancer [\#63](https://github.com/NodeFactoryIo/vedran/pull/63) ([MakMuftic](https://github.com/MakMuftic))
+
+### Fix
+
+### Changed
+
 ## [v0.1.1]((https://github.com/NodeFactoryIo/vedran/tree/v0.1.1))
 
 [Full Changelog](https://github.com/NodeFactoryIo/vedran/compare/v.0.1.0...v.0.1.1)